# backend/api/openapi/rime-v1.yaml
openapi: 3.1.0
info:
  title: Rime Native API
  version: 1.0.0
  description: Native API for catalog search and playback resolution.
servers:
  - url: /api/v1
security:
  - cookieAuth: []
  - bearerAuth: []
paths:
...
  /auth/setup:
    post:
      operationId: setupInitialAdmin
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/SetupRequest' }
      responses:
        '201':
          description: Initial administrator created and signed in
          content:
            application/json:
              schema: { $ref: '#/components/schemas/User' }
        '400': { $ref: '#/components/responses/Problem' }
        '409': { $ref: '#/components/responses/Problem' }
  /auth/login:
    post:
      operationId: login
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/LoginRequest' }
      responses:
        '200':
          description: Signed in user
          content:
            application/json:
              schema: { $ref: '#/components/schemas/User' }
        '401': { $ref: '#/components/responses/Problem' }
        '429': { $ref: '#/components/responses/Problem' }
  /auth/token/setup:
    post:
      operationId: setupInitialAdminToken
      security: []
      parameters:
        - $ref: '#/components/parameters/NativeClient'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/SetupRequest' }
      responses:
        '201':
          description: Initial administrator and native client session created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TokenSession' }
        '400': { $ref: '#/components/responses/Problem' }
        '409': { $ref: '#/components/responses/Problem' }
  /auth/token:
    post:
      operationId: createTokenSession
      security: []
      parameters:
        - $ref: '#/components/parameters/NativeClient'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/LoginRequest' }
      responses:
        '200':
          description: Native client session created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TokenSession' }
        '401': { $ref: '#/components/responses/Problem' }
        '429': { $ref: '#/components/responses/Problem' }
...
  /me/password:
    patch:
      operationId: changeOwnPassword
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [currentPassword, newPassword]
              properties:
                currentPassword: { type: string, format: password }
                newPassword: { type: string, format: password, minLength: 8, maxLength: 128 }
      responses:
        '204': { description: Password changed; all sessions revoked }
        '400': { $ref: '#/components/responses/Problem' }
        '401': { $ref: '#/components/responses/Problem' }
...
  /me/playlists:
    get:
      operationId: listOwnPlaylists
      parameters:
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 50, default: 10 }
        - name: cursor
          in: query
          schema: { type: string }
      responses:
        '200':
          description: Current user's playlists
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PlaylistPage' }
    post:
      operationId: createOwnPlaylist
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name: { type: string, minLength: 1, maxLength: 80 }
      responses:
        '201':
          description: Playlist created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Playlist' }
...
  /me/favorites/tracks/{trackId}:
    parameters:
      - $ref: '#/components/parameters/TrackId'
    get:
      operationId: getFavoriteStatus
      responses:
        '200':
          description: Favorite state for a track
          content:
            application/json:
              schema:
                type: object
                required: [favorite]
                properties:
                  favorite: { type: boolean }
    put:
      operationId: addFavorite
      responses:
        '204': { description: Track is a favorite }
    delete:
      operationId: removeFavorite
      responses:
        '204': { description: Track is not a favorite }
...
  /me/favorites/albums/{albumId}:
    parameters:
      - name: albumId
        in: path
        required: true
        schema: { type: string }
    get:
      operationId: getFavoriteAlbumStatus
      responses:
        '200':
          description: 当前用户对专辑的直接喜欢状态
          content:
            application/json:
              schema:
                type: object
                required: [favorite]
                properties:
                  favorite: { type: boolean }
    put:
      operationId: addFavoriteAlbum
      responses:
        '204': { description: Album is a favorite }
        '404': { $ref: '#/components/responses/Problem' }
    delete:
      operationId: removeFavoriteAlbum
      responses:
        '204': { description: Album is not a favorite }
  /admin/users:
    get:
      operationId: listUsers
      responses:
        '200':
          description: All users; administrator only
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items: { type: array, items: { $ref: '#/components/schemas/User' } }
        '403': { $ref: '#/components/responses/Problem' }
    post:
      operationId: createUser
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateUserRequest' }
      responses:
        '201':
          description: User created with a favorites playlist
          content:
            application/json:
              schema: { $ref: '#/components/schemas/User' }
        '403': { $ref: '#/components/responses/Problem' }
        '409': { $ref: '#/components/responses/Problem' }
  /admin/users/{userId}:
    patch:
      operationId: updateUser
      parameters:
        - $ref: '#/components/parameters/UserId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                displayName: { type: string, minLength: 1, maxLength: 40 }
                role: { type: string, enum: [admin, user] }
                disabled: { type: boolean }
      responses:
        '200':
          description: User updated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/User' }
        '403': { $ref: '#/components/responses/Problem' }
        '409': { $ref: '#/components/responses/Problem' }
...
  /playback/sessions:
    post:
      operationId: createPlaybackSession
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePlaybackSession'
      responses:
        '201':
          description: Resolved playback session
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlaybackSession'
        '404':
          $ref: '#/components/responses/Problem'
        '409':
          $ref: '#/components/responses/Problem'
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: rime_session
    bearerAuth:
      type: http
      scheme: bearer
  parameters:
    NativeClient:
      name: X-Rime-Client
      in: header
      required: true
      schema: { type: string, enum: [tauri] }
...
    TrackId:
      name: trackId
      in: path
      required: true
      schema: { type: string }
    UserId:
      name: userId
      in: path
      required: true
      schema: { type: string }
  responses:
    Problem:
      description: RFC 9457 problem details
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
  schemas:
    User:
      type: object
      required: [id, username, displayName, role, disabled, mustChangePassword, createdAt]
      properties:
        id: { type: string }
        username: { type: string }
        displayName: { type: string }
        role: { type: string, enum: [admin, user] }
        disabled: { type: boolean }
        mustChangePassword: { type: boolean }
        createdAt: { type: string, format: date-time }
        lastLoginAt: { type: string, format: date-time }
...
    TokenSession:
      type: object
      required: [accessToken, expiresAt, user]
      properties:
        accessToken: { type: string }
        expiresAt: { type: string, format: date-time }
        user: { $ref: '#/components/schemas/User' }
    SetupRequest:
      type: object
      required: [username, password]
      properties:
        username: { type: string, minLength: 3, maxLength: 32 }
        displayName: { type: string, maxLength: 40 }
        password: { type: string, format: password, minLength: 8, maxLength: 128 }
    LoginRequest:
      type: object
      required: [username, password]
      properties:
        username: { type: string }
        password: { type: string, format: password }
    CreateUserRequest:
      type: object
      required: [username, password, role]
      properties:
        username: { type: string, minLength: 3, maxLength: 32 }
        displayName: { type: string, maxLength: 40 }
        password: { type: string, format: password, minLength: 8, maxLength: 128 }
        role: { type: string, enum: [admin, user] }
    Playlist:
      type: object
      required: [id, name, kind, trackCount, createdAt, updatedAt]
      properties:
        id: { type: string }
        name: { type: string }
        kind: { type: string, enum: [favorites, custom] }
        trackCount: { type: integer, minimum: 0 }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
...
    PlaylistPage:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items: { $ref: '#/components/schemas/Playlist' }
        nextCursor: { type: string }
...
    ArtistRef:
      type: object
      required: [id, name]
      properties:
        id: { type: string }
        name: { type: string }
        role: { type: string }
    AlbumRef:
      type: object
      required: [id, title]
      properties:
        id: { type: string }
        title: { type: string }
...
    Track:
      type: object
      required: [id, title, album, artists, durationMs, available]
      properties:
        id: { type: string }
        title: { type: string }
        album: { $ref: '#/components/schemas/AlbumRef' }
        artists:
          type: array
          items: { $ref: '#/components/schemas/ArtistRef' }
        durationMs: { type: integer, format: int64 }
        discNumber: { type: integer }
        trackNumber: { type: integer }
        artworkId: { type: string }
        artworkFocus:
          type: object
          required: [x, y]
          properties:
            x: { type: number, minimum: 0, maximum: 1 }
            y: { type: number, minimum: 0, maximum: 1 }
        available: { type: boolean }
...
    MediaFormat:
      type: object
      required: [container]
      properties:
        container: { type: string }
        codec: { type: string }
    CreatePlaybackSession:
      type: object
      required: [trackId, playerId, capabilities]
      properties:
        trackId: { type: string }
        startPositionMs: { type: integer, format: int64, minimum: 0 }
        playerId: { type: string }
        capabilities:
          type: object
          required: [formats, supportsByteRange]
          properties:
            formats:
              type: array
              items: { $ref: '#/components/schemas/MediaFormat' }
            supportsByteRange: { type: boolean }
    PlaybackSource:
      type: object
      required: [kind, href, contentType, container, seekMethod]
      properties:
        kind: { type: string, enum: [direct, transcode] }
        href: { type: string }
        contentType: { type: string }
        container: { type: string }
        codec: { type: string }
        bitrateKbps:
          type: integer
          minimum: 1
          description: Effective stream bitrate; for direct playback this is the source file bitrate.
        seekMethod: { type: string, enum: [byteRange, timeOffset, none] }
    PlaybackSession:
      type: object
      required: [sessionId, track, source, expiresAt]
      properties:
        sessionId: { type: string }
        track: { $ref: '#/components/schemas/Track' }
        source: { $ref: '#/components/schemas/PlaybackSource' }
        expiresAt: { type: string, format: date-time }
...
    Problem:
      type: object
      required: [type, title, status, detail, instance, code]
      properties:
        type: { type: string }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        instance: { type: string }
        code: { type: string }
        traceId: { type: string }